+++ R/dashbioSpeck.R
@@ -1,8 +1,8 @@
 # AUTO GENERATED FILE - DO NOT EDIT
 
-dashbioSpeck <- function(id=NULL, data=NULL, loading_state=NULL, presetView=NULL, scrollZoom=NULL, view=NULL) {
+dashbioSpeck <- function(id=NULL, data=NULL, loading_state=NULL, presetView=NULL, scrollZoom=NULL, showLegend=NULL, style=NULL, view=NULL) {
     
-    props <- list(id=id, data=data, loading_state=loading_state, presetView=presetView, scrollZoom=scrollZoom, view=view)
+    props <- list(id=id, data=data, loading_state=loading_state, presetView=presetView, scrollZoom=scrollZoom, showLegend=showLegend, style=style, view=view)
     if (length(props) > 0) {
         props <- props[!vapply(props, is.null, logical(1))]
     }
@@ -10,7 +10,7 @@ dashbioSpeck <- function(id=NULL, data=NULL, loading_state=NULL, presetView=NULL
         props = props,
         type = 'Speck',
         namespace = 'dash_bio',
-        propNames = c('id', 'data', 'loading_state', 'presetView', 'scrollZoom', 'view'),
+        propNames = c('id', 'data', 'loading_state', 'presetView', 'scrollZoom', 'showLegend', 'style', 'view'),
         package = 'dashBio'
         )